dodana ładna strutkrua plików m.i.n graph.h i utils.h
