+++ graph.h
@@ -1,0 +1,26 @@
+#ifndef GRAPH_H
+#define GRAPH_H
+
+#include <stdio.h>
+#include <stdlib.h>
+
+// Structure for adjacency list node
+typedef struct Node {
+    int vertex;
+    struct Node* next;
+} Node;
+
+// Graph structure
+typedef struct Graph {
+    int numVertices;
+    Node** adjLists;
+} Graph;
+
+// Function prototypes
+Node* createNode(int v);
+Graph* createGraph(int vertices);
+void addEdge(Graph* graph, int src, int dest);
+void printGraph(Graph* graph);
+Graph* loadGraph(FILE* file);
+
+#endif
