+++ utils.h
@@ -1,0 +1,13 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+#include <ctype.h>
+
+// Function prototypes
+int how_many_digits(const char *line);
+void read_digits(char *line4, int *connections);
+
+#endif
